:sparkles: Enable to run `patch()` with an existing socket
