+++ browser/websocket/mod.ts
@@ -1,3 +1,4 @@
 export * from "./room.ts";
 export * from "./shortcuts.ts";
+export * from "./patch.ts";
 export * from "./stream.ts";
